+++ src/jrd/PBGen.cpp
@@ -49,7 +49,7 @@ void PBGen::putParameter(UCHAR parameter, int value)
 void PBGen::putParameter(UCHAR parameter, const TEXT* string)
 {
 	appendUCHAR(parameter);
-	int length = strlen (string);
+	int length = (int) strlen (string);
 	appendUCHAR(length);
 	appendCharacters(length, string);
 }
